[WIP - Checkpoint] Switching accounts if necessary when loading latest note works
